+++ test/PlayerUtil.hs
@@ -20,14 +20,22 @@ getPlayerPosition :: Player -> Position2D
 getPlayerPosition player = getCirclePosition $ playerToCircle player
 
 getPlayerAngle :: Player -> Angle2D
-getPlayerAngle (Player _ _ (Aim2D _ _ angle) _ _) = angle
+getPlayerAngle (Player _ _ (Gun (Aim2D _ _ angle) _ _) _) = angle
 
 getReloadTime :: Player -> ReloadTime
-getReloadTime (Player _ _ _ reloadTime _) = reloadTime
+getReloadTime (Player _ _ (Gun _ reloadTime _) _) = reloadTime
+
+getGun :: Player -> Gun
+getGun (Player _ _ gun _) = gun
+
+setReloadTime :: ReloadTime -> Player -> Player
+setReloadTime reloadTime player =
+    let Player _ _ (Gun aim _ state) _ = player
+    in  setGun (Gun aim reloadTime state) player
 
 setPlayerVelocity :: Velocity2D -> Player -> Player
-setPlayerVelocity velocity (Player circle _ aim reloadTime joystickId) =
-    Player circle velocity aim reloadTime joystickId
+setPlayerVelocity velocity (Player circle _ gun joystickId) =
+    Player circle velocity gun joystickId
 
 getRequiredStickPosition :: Vector1D -> Time -> Integer
 getRequiredStickPosition distance time =