+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Limited amount of shots per second for each player
+* Players keep firing while trigger is pressed
 
 ## [0.1.1] - 2019-12-21
 ### Fixed